refact:frontend components
