+++ frontend/src/lib/supabase.ts
@@ -1,0 +1,14 @@
+// lib/supabase.ts
+
+import { createClient } from "@supabase/supabase-js";
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!;
+const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!;
+
+if (!supabaseUrl || !supabaseAnonKey) {
+  throw new Error(
+    "Supabase URL 或 Anon Key 缺失。请确保它们在您的环境变量中设置。"
+  );
+}
+
+export const supabase = createClient(supabaseUrl, supabaseAnonKey);
