--- frontend/src/lib/supabaseClient.ts
@@ -1,7 +1,0 @@
-// lib/supabaseClient.ts
-import { createBrowserClient } from "@supabase/ssr";
-
-export const supabase = createBrowserClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-);
