+++ frontend/src/types/index.ts
@@ -1,0 +1,76 @@
+import type { PostgrestError } from "@supabase/supabase-js";
+
+export interface ArticleItem {
+  id: number;
+  title: string;
+  url: string;
+  published_date: string;
+  tags: string[];
+  summary: string | null;
+}
+
+export interface TelegramUser {
+  id: number;
+  first_name: string;
+  last_name?: string;
+  username?: string;
+  photo_url?: string;
+  auth_date: number;
+  hash: string;
+}
+
+export interface TelegramWebApp {
+  version: string;
+  platform: string;
+  colorScheme: "light" | "dark";
+  themeParams: {
+    bg_color: string;
+    text_color: string;
+    hint_color: string;
+    link_color: string;
+    button_color: string;
+    button_text_color: string;
+    secondary_bg_color?: string;
+  };
+  initDataUnsafe: {
+    query_id?: string;
+    user?: {
+      id: number;
+      first_name: string;
+      last_name?: string;
+      username?: string;
+      language_code?: string;
+      photo_url?: string;
+    };
+    auth_date: number;
+    hash: string;
+  };
+  initData: string;
+  ready: () => void;
+  expand: () => void;
+  close: () => void;
+  MainButton: {
+    text: string;
+    color: string;
+    textColor: string;
+    isVisible: boolean;
+    isActive: boolean;
+    isProgressVisible: boolean;
+    setText: (text: string) => void;
+    onClick: (callback: () => void) => void;
+    show: () => void;
+    hide: () => void;
+    enable: () => void;
+    disable: () => void;
+    showProgress: (leaveActive: boolean) => void;
+    hideProgress: () => void;
+  };
+}
+
+declare global {
+  interface Window {
+    Telegram?: {
+      WebApp: TelegramWebApp;
+    };
+  }
+}
